+++ plugins/kernel_v2/config/vm.rb
@@ -660,6 +660,15 @@ module VagrantPlugins
           end
         end
 
+        # Validate sub-VMs if there are any
+        @__defined_vms.each do |name, _|
+          if name =~ /[\[\]\{\}]/
+            errors["vm"] << I18n.t(
+              "vagrant.config.vm.name_invalid",
+              name: name)
+          end
+        end
+
         errors
       end
     end